requirements/base.txt
# Base will provide the common requirements among all environments (development, live, staging, etc).

# Django
Django==2.0.6

# Djongo - Driver for allowing Django to use MongoDB
djongo==1.2.29


# Enviroment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.8.2

# Django REST Framework JWT token
djangorestframework-jwt==1.11.0

# Django REST Framework JSON.API Format
djangorestframework-jsonapi==2.5.0

# Timezone
pytz==2018.4

# Windows's Interface to access  CA and CRL certificates.
wincertstore==0.2

# CORS
django-cors-headers==2.4.0
